refactor: mv coalesceWalletState to wallet-cli
